+++ editor/src/ek/editor/assets/dynamic_atlas_editor_asset.hpp
@@ -23,7 +23,8 @@ public:
     void save() override;
 
 private:
-
+    bool alphaMap = false;
+    bool mipmaps = false;
 };